requirements.txt
ccxt>=4.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
urllib3>=2.0.0
Flask>=3.0.0
pandas-ta==0.3.14b0
requests>=2.31.0
gunicorn>=21.0.0
